tidy: Explicitly convert boolean<->int when r/w QDomElement attributes.
The clang-tidy "implicit boolean conversion" check pointed out a
couple of places where booleans were being implicitly converted to
integers when setting attributes on a QDomElement, or implicitly being
converted to boolean when reading attributes from a QDomELement.  Be
explicit about these conversion.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html

+++ mythplugins/mytharchive/mytharchive/exportnative.cpp
@@ -403,11 +403,11 @@ void ExportNative::createConfigFile(const QString &filename)
 
     // add the options to the xml file
     QDomElement options = doc.createElement("options");
-    options.setAttribute("createiso", m_bCreateISO);
-    options.setAttribute("doburn", m_bDoBurn);
+    options.setAttribute("createiso", static_cast<int>(m_bCreateISO));
+    options.setAttribute("doburn", static_cast<int>(m_bDoBurn));
     options.setAttribute("mediatype", m_archiveDestination.type);
     options.setAttribute("dvdrsize", (qint64)m_archiveDestination.freeSpace);
-    options.setAttribute("erasedvdrw", m_bEraseDvdRw);
+    options.setAttribute("erasedvdrw", static_cast<int>(m_bEraseDvdRw));
     options.setAttribute("savedirectory", m_saveFilename);
     job.appendChild(options);
